fix(platform): address PR review — add tests, remove unused import, lift shared component
- Remove unused `logger` import from ask_question.py
- Add colocated ask_question_test.py with 3 tests covering options,
  no-options, and keyword-only cases
- Move ClarificationQuestionsCard to shared copilot/components/ location
  so AskQuestion imports from there instead of CreateAgent internals
